+++ app/src/main/res/layout/attraction_details_header.xml
