+++ app/src/main/res/layout/attraction_details_info_item.xml
